requirements.txt
openai
streamlit
gitpython
langchain
langchain-text-splitters
python-dotenv
tiktoken
